requirements.txt
anthropic>=0.42.0
feedparser>=6.0.0
trafilatura>=1.12.0
pyyaml>=6.0
python-frontmatter>=1.1.0
tenacity>=9.0.0
